requirements.txt
pynacl==1.6.2
base58==2.1.1
requests==2.32.5
eth-account==0.13.7